Thêm chức năng review phía client và admin

+++ nhom10/src/main/java/com/haui/service/UserService.java
@@ -36,11 +36,11 @@ public class UserService {
         return userRepository.findAll(spec, pageable);
     }
 
-    public User findById(Long id) {
+    public User findById(int id) {
         return userRepository.findById(id).get();
     }
 
-    public void delete(Long id) {
+    public void delete(int id) {
         userRepository.deleteById(id);
     }
